+++ lib/templateName.js
@@ -1,0 +1,3 @@
+module.exports = function (name) {
+    return 'tmpl_' + name.replace(/[^A-Za-z0-9]/g, '_');
+};
